+++ mapie/risk_control/__init__.py
@@ -1,14 +1,17 @@
 from .binary_classification import BinaryClassificationController
 from .multi_label_classification import MultiLabelClassificationController
-from .semantic_segmentation import SemanticSegmentationController
 from .risks import (
     BinaryClassificationRisk,
+    abstention_rate,
     accuracy,
     false_positive_rate,
+    negative_predictive_value,
+    positive_predictive_value,
     precision,
     predicted_positive_fraction,
     recall,
 )
+from .semantic_segmentation import SemanticSegmentationController
 
 __all__ = [
     "MultiLabelClassificationController",
@@ -20,4 +23,7 @@ __all__ = [
     "precision",
     "recall",
     "predicted_positive_fraction",
+    "positive_predictive_value",
+    "negative_predictive_value",
+    "abstention_rate",
 ]