+++ HISTORY.rst
@@ -16,6 +16,7 @@ History
 * Fix data leakage for time series tutorials
 * Improve documentation display (increase width and improve API table)
 * fix bug with CRC and RCPS where the computed lambda was not the best
+* add example of LLM as a judge with abstention
 
 1.2.0 (2025-11-17)
 ------------------